+++ pad/editor/src/utils.rs
@@ -410,7 +410,7 @@ pub fn update_ctrl(event: &Event) {
         .set_class_name(if os_ctrl(event) { "ctrl-pressed" } else { "" });
 }
 
-fn build_code_lines(id: &str, code: &str, hidden: &str) -> CodeLines {
+fn build_code_lines(code: &str, hidden: &str) -> CodeLines {
     let mut lines = CodeLines {
         frags: vec![Vec::new()],
     };
@@ -457,7 +457,7 @@ fn build_code_lines(id: &str, code: &str, hidden: &str) -> CodeLines {
     };
 
     let mut end = 0;
-    let spans = Spans::with_backend(&full_code, WebBackend::new(id, &full_code));
+    let spans = Spans::with_backend(&full_code, WebBackend::new());
     for span in spans.spans {
         let kind = span.value;
         let span = span.span;
@@ -555,7 +555,7 @@ fn build_code_lines(id: &str, code: &str, hidden: &str) -> CodeLines {
     lines
 }
 
-pub fn gen_code_view(id: &str, code: &str, hidden: &str) -> View {
+pub fn gen_code_view(code: &str, hidden: &str) -> View {
     fn pair_aliases() -> HashMap<(Primitive, Primitive), &'static str> {
         use Primitive::*;
         [
@@ -663,7 +663,7 @@ pub fn gen_code_view(id: &str, code: &str, hidden: &str) -> View {
     }
 
     // logging::log!("gen_code_view({code:?})");
-    let CodeLines { frags } = build_code_lines(id, code, hidden);
+    let CodeLines { frags } = build_code_lines(code, hidden);
     let mut line_views = Vec::new();
     let gayness = get_gayness();
     for line in frags {
@@ -1085,14 +1085,14 @@ pub fn gen_code_view(id: &str, code: &str, hidden: &str) -> View {
     line_views.into_view()
 }
 
-fn init_rt(id: &str, code: &str) -> Uiua {
-    Uiua::with_backend(WebBackend::new(id, code))
+fn init_rt() -> Uiua {
+    Uiua::with_backend(WebBackend::new())
         .with_execution_limit(Duration::from_secs_f64(get_execution_limit()))
         .with_recursion_limit(50)
 }
 
-fn just_values(id: &str, code: &str) -> UiuaResult<Vec<Value>> {
-    let mut rt = init_rt(id, code);
+fn just_values(code: &str) -> UiuaResult<Vec<Value>> {
+    let mut rt = init_rt();
     rt.run_str(code)?;
     Ok(rt.take_stack())
 }
@@ -1114,10 +1114,11 @@ impl State {
             format!("{}\n{code}", self.hidden)
         };
         if let Some(chal) = &self.challenge {
-            let mut example = run_code_single(
-                &self.code_id,
-                &challenge_code(&chal.intended_answer, &chal.example, chal.flip),
-            )
+            let mut example = run_code_single(&challenge_code(
+                &chal.intended_answer,
+                &chal.example,
+                chal.flip,
+            ))
             .0;
             example.insert(
                 0,
@@ -1126,21 +1127,17 @@ impl State {
             let mut output_sections = vec![example];
             let mut correct = true;
             for (i, test) in chal.tests.iter().enumerate() {
-                let answer = || {
-                    just_values(
-                        &self.code_id,
-                        &challenge_code(&chal.intended_answer, test, chal.flip),
-                    )
-                };
+                let answer =
+                    || just_values(&challenge_code(&chal.intended_answer, test, chal.flip));
                 let user_input = challenge_code(&full_code, test, chal.flip);
-                let user_output = || just_values(&self.code_id, &user_input);
+                let user_output = || just_values(&user_input);
                 correct = correct
                     && match (answer(), user_output()) {
                         (Ok(answer), Ok(users)) => answer == users,
                         (Err(answer), Err(users)) => answer.to_string() == users.to_string(),
                         _ => false,
                     };
-                let mut output = run_code_single(&self.code_id, &user_input).0;
+                let mut output = run_code_single(&user_input).0;
                 output.insert(
                     0,
                     vec![OutputItem::Faint(format!("Test {}: {test}", i + 1))],
@@ -1163,7 +1160,7 @@ impl State {
             }
             output
         } else {
-            let (output, error) = run_code_single(&self.code_id, &full_code);
+            let (output, error) = run_code_single(&full_code);
             self.loading_module = false;
             if let Some(error) = error
                 && error.to_string().contains("Waiting for module")
@@ -1176,18 +1173,17 @@ impl State {
 }
 
 #[allow(clippy::mutable_key_type)]
-fn run_code_single(id: &str, code: &str) -> (Vec<Vec<OutputItem>>, Option<UiuaError>) {
+fn run_code_single(code: &str) -> (Vec<Vec<OutputItem>>, Option<UiuaError>) {
     // Run
-    let mut rt = init_rt(id, code);
+    let mut rt = init_rt();
     let mut error = None;
-    let mut comp = Compiler::with_backend(WebBackend::new(id, code));
+    let mut comp = Compiler::with_backend(WebBackend::new());
     let comp_backend;
     let res = comp.load_str(code).map(|comp| rt.run_compiler(comp));
     let (mut value_lines, io) = match res {
         Ok(Ok(())) => {
             let stack = rt.take_stack_lines();
             let backend = rt.downcast_backend::<WebBackend>().unwrap();
-            backend.finish();
             (stack, backend)
         }
         Ok(Err(e)) if matches!(*e.kind, UiuaErrorKind::Interrupted) => (